👌✨ Improve UX/UI in Home.razor and add flash message functionality
Key changes:
- Implemented flash messages to indicate when no aliases are found
- Replaced standard input with InputNumber and InputText
- Enhanced style and UI of Search form

The primary changes were the addition of flash message functionality to indicate when no aliases are found in the search. This was done to give clearer feedback to the user. The input fields were also replaced with Blazor InputNumber and InputText to standardize input fields across the application. Lastly, the aesthetic and user interface of the Search form were enhanced to improve the user experience.